requirements/dev.txt
-r base.txt
flake8~=5.0
pytest~=7.3
pytest-cookies~=0.7
pytest-instafail~=0.5
pyyaml~=6.0
mypy~=1.2
bandit~=1.7